+++ internal/constants/constants.go
@@ -89,6 +89,8 @@ const (
 	//
 
 	KubernetesJoinTokenTTL = 15 * time.Minute
+	ConstellationNamespace = "kube-system"
+	JoinConfigMap          = "join-config"
 
 	//
 	// Helm.