+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GCP service account can now be managed manually.
 - CoreOS images are publicly available for Azure.
 - GCP: Support for higher end N2D standard (128 & 224 vCPUs), *high-mem* and *high-cpu* VMs
+- Add `constellation upgrade` to update node images in Constellation.
 
 ### Changed
 <!-- For changes in existing functionality.  -->